typescriptanaization
