+++ packages/core/src/interfaces/has-page-id.ts
@@ -1,0 +1,3 @@
+export type HasPageId = {
+  pageId: string,
+};
